chore: configure storybook paths and fonts
