+++ .storybook/main.js
@@ -1,3 +1,5 @@
+const { imageConfigDefault } = require("next/dist/shared/lib/image-config");
+const path = require("path");
 module.exports = {
   "stories": [
     "../src/**/*.stories.mdx",
@@ -11,5 +13,13 @@ module.exports = {
   "framework": "@storybook/react",
   "core": {
     "builder": "@storybook/builder-webpack5"
+  }, 
+  staticDirs: ['../public'],
+  webpackFinal: async (config, { configType }) => {
+    config.resolve.modules = [
+      ...(config.resolve.modules || []),
+      path.resolve('./src'),
+    ];
+    return config;
   }
 }
