+++ .storybook/preview-head.html
@@ -1,0 +1,11 @@
+<style>
+@font-face {
+    font-family: "PingFang";
+    src: url('fonts/PingFang SC Regular.eot'),
+      url('fonts/PingFang SC Regular.eot?#') format('embedded-opentype'),
+      url('fonts/PingFang SC Regular.woff') format('woff'),
+      url('fonts/PingFang SC Regular.ttf') format('truetype');
+    font-weight: normal;
+    font-style: normal;
+  }
+</style>
