+++ pages/_document.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import NextDocument, { Html, Head, Main, NextScript } from 'next/document';
-import { getCssText } from '../stitches.config';
+import { getCssText } from 'themes';
 
 export default class Document extends NextDocument {
+ 
   render() {
     return (
       <Html lang="en">
