--- src/components/Button/__tests__/Button.spec.tsx
@@ -1,9 +1,0 @@
-import {render} from '@testing-library/react'
-import { Button } from "../";
-
-describe('First test', () => {
-    it('should exist', () => {
-        const {getByText} = render(<Button>Funcionou</Button>)
-        expect(getByText(/Funcionou/)).toBeTruthy()
-    });    
-});
