--- src/components/Button/styles.ts
@@ -1,5 +1,0 @@
-import { styled } from '../../../stitches.config';
-export const Buttom = styled('button', {
-    fontFamily: '$system',
-    color: 'green',
-  });
