+++ src/components/Button/styles/Button.styles.ts
@@ -1,0 +1,9 @@
+import { styled } from 'themes';
+export const Buttom = styled('button', {
+    fontFamily: '$system',
+    lineHeight: '1.15',
+    fontSize: '5rem',
+    color: '$default',
+    padding: '0 2rem',
+    fontWeight: 'bold'
+  });
