+++ src/components/Button/types/Button.type.ts
@@ -1,0 +1,3 @@
+export type Props = {
+    children: JSX.Element | string,
+};
