+++ src/components/menu/Item/styles.ts
@@ -1,0 +1,5 @@
+import { styled } from 'themes/main.theme';
+export const Buttom = styled('div', {
+    fontFamily: '$system',
+    color: 'red',
+  });
