+++ src/themes/main.theme.ts
@@ -1,0 +1,19 @@
+import { createStitches } from '@stitches/react';
+
+export const { styled, getCssText } = createStitches({
+  theme: {
+    fonts: {
+      system: 'PingFang',
+    },
+    colors: {
+      default: 'green', 
+      hiContrast: 'hsl(206,10%,5%)',
+      loContrast: 'white',
+    },
+    fontSizes: {
+      1: '13px',
+      2: '15px',
+      3: '17px',
+    },
+  },
+});
